# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentazione API con Swagger",
        "title": "API Server Go",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Risponde \"status healthy\" sulla root",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "status healthy",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/llm": {
            "post": {
                "description": "Invia un prompt all'LLM e ritorna la risposta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "Ask LLM",
                "parameters": [
                    {
                        "description": "Prompt body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LLMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.LLMResponse"
                        }
                    }
                }
            }
        },
        "/naruto/chat": {
            "post": {
                "description": "Cerca su Dattebayo e risponde in italiano pulito via OpenAI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "naruto"
                ],
                "summary": "Chat Naruto (Dattebayo + OpenAI)",
                "parameters": [
                    {
                        "description": "Messaggio utente",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.NarutoChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.NarutoChatResponse"
                        }
                    }
                }
            }
        },
        "/saluta": {
            "get": {
                "description": "Risponde \"ciao\" o \"ciao {nome}\"",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Saluta il server",
                "responses": {
                    "200": {
                        "description": "ciao\"  \"ciao {nome}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/saluta/con-bestemmia": {
            "get": {
                "description": "Risponde \"Porco Dio\" o \"ciao {nome} che voi Porco Dio\"",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Bestemmia il server",
                "responses": {
                    "200": {
                        "description": "Porco Dio\"  \"ciao {nome} che voi Porco Dio",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.LLMRequest": {
            "type": "object",
            "properties": {
                "prompt": {
                    "type": "string"
                }
            }
        },
        "main.LLMResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                }
            }
        },
        "main.NarutoChatRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.NarutoChatResponse": {
            "type": "object",
            "properties": {
                "reply": {
                    "type": "string"
                }
            }
        }
    }
}